Adds multhreaded version of pingr application
The application expects an argument to be passed. This argument is the
base port. Currently, the application probes the 10 following port of
the given port with each 1000 simple curl requests. This probing is done
in parallel, where each port is probed by a single thread. So currently
the application spawns 10 threads (for the 10 different ports) and then
tests them.
The behavior is not that flexible, both the attempts and the replicas
should be defined better, this should be done by a argument parsing
library.